// src/components/client/GLRadarChart/types.ts
export interface RadarChartStyle {
  backgroundColor?: [number, number, number, number];
  lineColor?: [number, number, number, number];
  outlineColor?: [number, number, number, number];
  dataColor?: [number, number, number, number];
  dataOutlineColor?: [number, number, number, number];
  centerPointColor?: [number, number, number, number];
}

export interface RadarLoadingStyle {
  backgroundColor?: [number, number, number, number];
  outlineColor?: [number, number, number, number];
}